refactor: Computer 추상클래스로 추상화
