+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.controller.BaseBallGame;
+import baseball.controller.NumberBaseBallGame;
 import baseball.strategy.ForbidDuplicationPolicy;
 import baseball.strategy.RandomNumberGeneratePolicy;
 import baseball.view.ConsoleInput;
@@ -11,7 +11,7 @@ public class Application {
     private final static int NUMBER_OF_DIGITS = 3;
 
     public static void main(String[] args) {
-        new BaseBallGame(
+        new NumberBaseBallGame(
                 new ConsoleInput(),
                 new ConsoleOutput()
         ).run(
